Created single file for all animation components.
